+++ Bufor.h
@@ -12,11 +12,11 @@ class Bufor{
 	public:
 		Bufor(const int size);
 		~Bufor();
-		void setRandom(double minDist, double maxDist, double minW, double maxW);
+		void setRandom(float minDist, float maxDist, float minW, float maxW);
 		Cialo * getData();
 		const int getN()const;
-		void next(double deltaTime);
-//		void threadWorker(int b1, int e1, double deltaTime);
+		void next(float deltaTime);
+//		void threadWorker(int b1, int e1, float deltaTime);
 //		friend void threadWorker(Bufor&, int,int);
 	private:
 		Cialo * data;
