+++ Cialo.h
@@ -8,21 +8,21 @@ class Cialo{
 	public:
 		Cialo();
 		Cialo(const Cialo & dane);
-		Cialo(const Wektor& pozycja, const double& masa);
+		Cialo(const Wektor& pozycja, const float& masa);
 		Cialo(const Zakres& zakPoz, const Zakres& zakMasa);
 		Wektor & getPoz();
 		Wektor & getPred();
 		Wektor & getSila();
 		void set(const Zakres& zakPoz, const Zakres& zakMasa);
-		const double & getMasa() const;
+		const float & getMasa() const;
 		// siła działająca na ciało pochodząca od grawitacji ciała drugiego
 		Wektor silaPomiedzy(const Cialo & drugie) const;
 		// aktualizacja pozycji i predkości na podstawie aktualnej siły
-		void aktualizuj(double deltaTime);
+		void aktualizuj(float deltaTime);
 		// zeruje siłę
 		void clearF();
 		std::mutex mut;
 	private:
 		Wektor poz, pred, sila;
-		double masa;
+		float masa;
 };
