+++ Wektor.h
@@ -5,25 +5,25 @@
 
 class Wektor {
 	public:
-		Wektor(const double & x, const double & y, const double & z);
+		Wektor(const float & x, const float & y, const float & z);
 		Wektor();
-		const double len() const;
-		Wektor operator*(const double & a) const;
-		Wektor operator/(const double & a) const;
-		friend Wektor operator*(const double & a, const Wektor & vec);
+		const float len() const;
+		Wektor operator*(const float & a) const;
+		Wektor operator/(const float & a) const;
+		friend Wektor operator*(const float & a, const Wektor & vec);
 		friend std::ostream & operator<<(std::ostream & o, const Wektor & vec);
 		Wektor operator+(const Wektor & vec) const;
 		Wektor operator-(const Wektor & vec) const;
 		void operator+=(const Wektor & vec);
 		void operator-=(const Wektor & vec);
-		void operator*=(const double & a);
-		double & getX();
-		double & getY();
-		double & getZ();
+		void operator*=(const float & a);
+		float & getX();
+		float & getY();
+		float & getZ();
 		void clear();
 	private:
-		double x, y, z;
+		float x, y, z;
 };
-Wektor operator*(const double & a, const Wektor & vec);
+Wektor operator*(const float & a, const Wektor & vec);
 std::ostream & operator<<(std::ostream & o, const Wektor & vec);
