+++ Zakres.h
@@ -3,18 +3,18 @@
 #include <random>
 class Zakres{
 	public:
-		Zakres(double min, double max): min(min), max(max){};
-		Zakres(double max): min(0), max(max){};
-		Zakres(): min(std::numeric_limits<double>::min()), max(std::numeric_limits<double>::max()){};
-		double getMax()const {return max;}
-		double getMin()const {return min;}
-		double getRandom()const {
+		Zakres(float min, float max): min(min), max(max){};
+		Zakres(float max): min(0), max(max){};
+		Zakres(): min(std::numeric_limits<float>::min()), max(std::numeric_limits<float>::max()){};
+		float getMax()const {return max;}
+		float getMin()const {return min;}
+		float getRandom()const {
 			std::random_device rd;
 			std::default_random_engine gen(rd());
-			std::uniform_real_distribution<double> dist(min, max);
+			std::uniform_real_distribution<float> dist(min, max);
 			return dist(gen);
 		}
 	private:
-		const double min, max;
+		const float min, max;
 
 };
